Fetch relevant events first draft

+++ overlore/eternum/types.py
@@ -1,6 +1,6 @@
 from typing import TypeAlias
 
-RealmPosition: TypeAlias = tuple[int, int]
+RealmPosition: TypeAlias = tuple[float, float]
 ResourceAmount: TypeAlias = dict[str, int]
 ResourceAmounts: TypeAlias = list[ResourceAmount]
 AttackingEntityIds: TypeAlias = list[int]
